Changed logout logic to not demand cookies and more

+++ src/main/kotlin/com/roimolam/project/dal/CatalogItemDAL.kt
@@ -1,6 +1,7 @@
 package com.roimolam.project.dal
 
 import com.roimolam.project.constants.CATALOG_ITEMS_PER_PAGE
+import com.roimolam.project.data.CatalogItemDeleteStatusWrapper
 import com.roimolam.project.data.CatalogItemIDWrapper
 import com.roimolam.project.data.entities.CatalogItemEntity
 import com.roimolam.project.data.CatalogItemsWrapper
@@ -85,9 +86,15 @@ class CatalogItemDAL (@PersistenceContext val entityManager:EntityManager,
         return query.resultList as List<CatalogItemEntity>
     }
 
-    fun deleteCatalogItem(@PathVariable id: Long) {
+    @Transactional(propagation = Propagation.REQUIRED)
+    fun deleteCatalogItem(@PathVariable id: Long): CatalogItemDeleteStatusWrapper {
+        println(id)
         entityManager.find(CatalogItemEntity::class.java, id).apply {
-            entityManager.remove(this)
+            if (this != null) {
+                entityManager.remove(this)
+            }
         }
+
+        return CatalogItemDeleteStatusWrapper("OK")
     }
 }
